// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        PaymentRequest: {
            /**
             * Format: uuid
             * @description Invoice ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            invoiceId: string;
            /**
             * @description Payment amount
             * @example 500
             */
            amount: number;
            /**
             * Format: date-time
             * @description Payment date
             */
            paymentDate?: string;
        };
        /** @description Payment information */
        PaymentResponse: {
            /**
             * Format: uuid
             * @description Payment ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            id?: string;
            /**
             * Format: uuid
             * @description Invoice ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            invoiceId?: string;
            /**
             * @description Payment amount
             * @example 500
             */
            amount?: number;
            /**
             * Format: date-time
             * @description Payment date
             */
            paymentDate?: string;
            /**
             * Format: date-time
             * @description Creation timestamp
             */
            createdAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}